ADDED: To Check Maximum Amoung Three Float

+++ com/generics/MaximumOfThree.java
@@ -2,18 +2,17 @@ package com.generics;
 
 public class MaximumOfThree {
 
-	private Integer x, y, z, max;
+	private Float x, y, z, max;
 
 	// Create Parameterize Constructor
-	public MaximumOfThree(Integer x, Integer y, Integer z) {
+	public MaximumOfThree(Float x, Float y, Float z) {
 		super();
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 
-	// Create Method For Find Maximum Of Three
-	public Integer testMaximum() {
+	public Float testMaximum() {
 		max = x;
 		if (y.compareTo(x) > 0)
 			max = y;
